Input, output and postprocess reorg
