+++ examples/gnn_fraud_detection_pipeline/run.py
@@ -24,9 +24,9 @@ from morpheus.config import Config
 from morpheus.config import CppConfig
 from morpheus.config import PipelineModes
 from morpheus.stages.general_stages import MonitorStage
-from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.postprocess.serialize import SerializeStage
-from morpheus.stages.output.to_file import WriteToFileStage
+from morpheus.stages.input.file_source_stage import FileSourceStage
+from morpheus.stages.postprocess.serialize_stage import SerializeStage
+from morpheus.stages.output.write_to_file_stage import WriteToFileStage
 from morpheus.pipeline.pipeline import LinearPipeline
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from morpheus.utils.logging import configure_logging
