+++ morpheus/io/utils.py
@@ -1,0 +1,39 @@
+# SPDX-FileCopyrightText: Copyright (c) 2021-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import typing
+
+import pandas as pd
+
+import cudf
+
+from morpheus._lib.file_types import FileTypes
+from morpheus._lib.file_types import determine_file_type
+
+
+def filter_null_data(x: typing.Union[cudf.DataFrame, pd.DataFrame]):
+    """
+    Filters out null row in a dataframe's 'data' column if it exists.
+
+    Parameters
+    ----------
+    x : typing.Union[cudf.DataFrame, pandas.DataFrame]
+        The dataframe to fix.
+    """
+
+    if ("data" not in x):
+        return x
+
+    return x[~x['data'].isna()]
