+++ morpheus/messages/multi_ae_message.py
@@ -36,7 +36,7 @@ from morpheus.pipeline.file_types import FileTypes
 from morpheus.pipeline.pipeline import MultiMessageStage
 from morpheus.pipeline.pipeline import StreamPair
 from morpheus.stages.inference.inference_ae import MultiInferenceAEMessage
-from morpheus.stages.input.from_cloudtrail import CloudTrailSourceStage
+from morpheus.stages.input.cloud_trail_source_stage import CloudTrailSourceStage
 
 logger = logging.getLogger(__name__)
