+++ morpheus/stages/preprocess/preprocess_ae_stage.py
@@ -37,7 +37,7 @@ from morpheus.pipeline.file_types import FileTypes
 from morpheus.pipeline.pipeline import MultiMessageStage
 from morpheus.pipeline.pipeline import StreamPair
 from morpheus.stages.inference.inference_ae import MultiInferenceAEMessage
-from morpheus.stages.input.from_cloudtrail import CloudTrailSourceStage
+from morpheus.stages.input.cloud_trail_source_stage import CloudTrailSourceStage
 from morpheus.stages.preprocess.preprocess_base_stage import PreprocessBaseStage
 
 logger = logging.getLogger(__name__)
