+++ tests/test_cli.py
@@ -34,15 +34,15 @@ from morpheus.stages.inference.inference_ae import AutoEncoderInferenceStage
 from morpheus.stages.inference.inference_identity import IdentityInferenceStage
 from morpheus.stages.inference.inference_pytorch import PyTorchInferenceStage
 from morpheus.stages.inference.inference_triton import TritonInferenceStage
-from morpheus.stages.input.from_cloudtrail import CloudTrailSourceStage
-from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.input.from_kafka import KafkaSourceStage
-from morpheus.stages.postprocess.serialize import SerializeStage
-from morpheus.stages.output.to_file import WriteToFileStage
-from morpheus.stages.output.to_kafka import WriteToKafkaStage
-from morpheus.stages.postprocess.validation import ValidationStage
-from morpheus.stages.postprocess.mlflow_drift import MLFlowDriftStage
-from morpheus.stages.postprocess.timeseries import TimeSeriesStage
+from morpheus.stages.input.cloud_trail_source_stage import CloudTrailSourceStage
+from morpheus.stages.input.file_source_stage import FileSourceStage
+from morpheus.stages.input.kafka_source_stage import KafkaSourceStage
+from morpheus.stages.postprocess.serialize_stage import SerializeStage
+from morpheus.stages.output.write_to_file_stage import WriteToFileStage
+from morpheus.stages.output.write_to_kafka_stage import WriteToKafkaStage
+from morpheus.stages.postprocess.validation_stage import ValidationStage
+from morpheus.stages.postprocess.ml_flow_drift_stage import MLFlowDriftStage
+from morpheus.stages.postprocess.timeseries_stage import TimeSeriesStage
 from morpheus.stages.preprocess.preprocess_ae_stage import PreprocessAEStage
 from morpheus.stages.preprocess.preprocess_ae_stage import TrainAEStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
