+++ tests/test_file_in_out_stage_pipe.py
@@ -21,9 +21,9 @@ import pytest
 
 from morpheus.pipeline import LinearPipeline
 from morpheus.pipeline.file_types import FileTypes
-from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.input.utils import read_file_to_df
-from morpheus.stages.output.to_file import WriteToFileStage
+from morpheus.stages.input.file_source_stage import FileSourceStage
+from morpheus.io.deserializers import read_file_to_df
+from morpheus.stages.output.write_to_file_stage import WriteToFileStage
 from utils import TEST_DIRS
