+++ tests/test_sid.py
@@ -25,10 +25,10 @@ from morpheus.pipeline import LinearPipeline
 from morpheus.stages.general_stages import AddClassificationsStage
 from morpheus.stages.general_stages import MonitorStage
 from morpheus.stages.inference.inference_triton import TritonInferenceStage
-from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.postprocess.serialize import SerializeStage
-from morpheus.stages.output.to_file import WriteToFileStage
-from morpheus.stages.postprocess.validation import ValidationStage
+from morpheus.stages.input.file_source_stage import FileSourceStage
+from morpheus.stages.postprocess.serialize_stage import SerializeStage
+from morpheus.stages.output.write_to_file_stage import WriteToFileStage
+from morpheus.stages.postprocess.validation_stage import ValidationStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from morpheus.stages.preprocess.preprocessing import PreprocessNLPStage
 from utils import TEST_DIRS
