+++ tests/utils.py
@@ -20,7 +20,7 @@ import os
 from morpheus.config import Config
 from morpheus.pipeline.file_types import FileTypes
 from morpheus.stages.inference import inference_stage
-from morpheus.stages.input.utils import read_file_to_df
+from morpheus.io.deserializers import read_file_to_df
 from morpheus.messages import MultiMessage
 from morpheus.messages import MultiResponseProbsMessage
 from morpheus.messages import ResponseMemoryProbs
